+++ Controllers/RemindersController.cs
@@ -1,14 +1,8 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using CalendarAPI.Data;
 using CalendarAPI.Dtos;
 using CalendarAPI.Interfaces;
 using CalendarAPI.Mappers;
-using CalendarAPI.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 
 namespace CalendarAPI.Controllers
 {
@@ -18,29 +12,48 @@ namespace CalendarAPI.Controllers
     {
         private readonly IEventRepository _eventRepository;
         private readonly IReminderRepository _reminderRepository;
-        public RemindersController(IEventRepository eventRepository, IReminderRepository reminderRepository){
+
+        public RemindersController(IEventRepository eventRepository, IReminderRepository reminderRepository)
+        {
             _eventRepository = eventRepository;
             _reminderRepository = reminderRepository;
         }
 
+        [Authorize]
         [HttpPost]
-        public async Task<IActionResult> CreateReminders([FromForm] CreateReminderDto dto){
-            Event _event = await _eventRepository.GetByIdAsync(dto.EventId);
-            Reminder reminder = ReminderMappers.RegisterReminder(dto, _event);
-            _event.Reminders.Add(reminder);
+        public async Task<IActionResult> CreateReminder([FromForm] CreateReminderDto dto)
+        {
+            var eventEntity = await _eventRepository.GetByIdAsync(dto.EventId);
+            if (eventEntity == null)
+            {
+                return NotFound("Event not found.");
+            }
+
+            var reminder = ReminderMappers.RegisterReminder(dto, eventEntity);
+
+            eventEntity.Reminders.Add(reminder);
             await _reminderRepository.AddAsync(reminder);
-            await _eventRepository.UpdateAsync(_event);
-            return Ok();
+            await _eventRepository.UpdateAsync(eventEntity);
+
+            return Ok("Reminder created successfully.");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteReminders([FromRoute] string id){
-            Reminder reminder = await _reminderRepository.GetByIdAsync(id);
-            Event _event = reminder.Event;
-            _event.Reminders.Remove(reminder);
-            await _eventRepository.UpdateAsync(_event);
+        public async Task<IActionResult> DeleteReminder([FromRoute] string id)
+        {
+            var reminder = await _reminderRepository.GetByIdAsync(id);
+            if (reminder == null)
+            {
+                return NotFound("Reminder not found.");
+            }
+
+            var eventEntity = reminder.Event;
+            eventEntity.Reminders.Remove(reminder);
+            await _eventRepository.UpdateAsync(eventEntity);
             await _reminderRepository.RemoveAsync(reminder);
-            return Ok();
+
+            return Ok("Reminder removed successfully.");
         }
     }
 }
